Animated Bullet + Overhaul to Shooting's FiringSource
ShipControlComponent now has an exposed FiringSource that is set in a method called InitShip() that's called upon either start (if the ship existed in the scene in the beginning) or upon spawn (otherwise).

Bullets now fires in the direction of FiringSource.right (so red arrow if you were to select it).